Fix deadlock on main worker loop
Deadlock was caused by race condition when taking the length of a channel
while workers were still writing to it.  New version uses a WaitGroup to
batch concurrent requests and reliably determine the number of outstanding
URLs to collect.

Made the custom http.Client a global for connection reuse.

Synchronize getAnchors and getInputs with a WaitGroup instead of a channel

Add read locks to fix race condition when reading from visited.URLS